# api/openapi.yaml
openapi: "3.1.0"
info:
  title: Valuz Agent API
  version: "0.3.0"
  description: Valuz Agent API contract — single source of truth
servers:
  - url: http://localhost:8000
    description: Local development
paths:
...
  /v1/notifications/history:
    get:
      summary: Resolved (dismissed / handled) notifications, newest first
      description: |
        The read-only history behind the drawer's History tab. Returns resolved
        rows only — the open set stays on GET /v1/notifications. Paginated by a
        created_at cursor: pass the last entry's created_at as `before` to load
        the next page.
      operationId: listNotificationHistory
      tags: [notifications]
      parameters:
        - name: limit
          in: query
          required: false
          schema: {type: integer, minimum: 1, maximum: 200, default: 50}
        - name: before
          in: query
          required: false
          schema:
            type: integer
            format: int64
            description: "Only entries with created_at strictly below this (ms cursor)."
      responses:
        "200":
          description: Resolved notifications page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationHistoryResponse"
...
  /v1/notifications/{notification_id}:read:
    post:
      summary: Mark one notification read
      operationId: markNotificationRead
      tags: [notifications]
      parameters:
        - name: notification_id
          in: path
          required: true
          schema: {type: string}
      responses:
        "200":
          description: OK
  /v1/notifications/{notification_id}:dismiss:
    post:
      summary: Dismiss (resolve) one notification
      operationId: dismissNotification
      tags: [notifications]
      parameters:
        - name: notification_id
          in: path
          required: true
          schema: {type: string}
      responses:
        "200":
          description: OK

  # ── Local backup (docs/design/client-local-backup.md) ──────────────
...
components:
...
  schemas:
...
    NotificationEntry:
      type: object
      required: [id, kind, title, body, action, urgency, payload, created_at]
      properties:
        id: {type: string}
        kind: {type: string, description: "question | task_failed | …"}
        title: {type: string}
        body: {type: string}
        route: {type: string, nullable: true}
        action: {type: string, description: "answer | resume | none"}
        urgency: {type: string, description: "actionable | info"}
        task_id: {type: string, nullable: true}
        project_id: {type: string, nullable: true}
        session_id: {type: string, nullable: true}
        pending_id: {type: string, nullable: true}
        payload: {type: object, additionalProperties: true}
        created_at: {type: integer, format: int64}
        read_at: {type: integer, format: int64, nullable: true}
        resolved_at: {type: integer, format: int64, nullable: true}
...
    NotificationHistoryResponse:
      type: object
      required: [entries]
      properties:
        entries:
          type: array
          items:
            $ref: "#/components/schemas/NotificationEntry"
        has_more:
          type: boolean
          description: "True when another page exists below the returned entries."
...